@@ -1,31 +1,32 @@
 {
-  "name": "microblog",
-  "version": "0.0.1",
-  "private": true,
+  "name": "micro-blog",
+  "version": "2.0.0",
+  "description": "Micro blog",
+  "main": "index.js",
   "scripts": {
-    "start": "node ./bin/www"
+    "start": "node bin/www.mjs"
   },
+  "author": "",
+  "license": "ISC",
   "dependencies": {
-    "body-parser": "^1.17.2",
-    "bson": "^1.1.4",
-    "connect-flash": "~0.1.1",
-    "connect-mongo": "^0.8.2",
-    "cookie-parser": "^1.4.3",
-    "debug": "^2.6.9",
-    "ejs": "^2.5.6",
-    "eslint-plugin-import": "^2.3.0",
-    "eslint-plugin-node": "^5.0.0",
-    "express": "^4.15.3",
-    "express-partials": "~0.3.0",
-    "express-session": "^1.15.3",
-    "mongodb": "^3.1.13",
-    "morgan": "^1.9.1",
-    "serve-favicon": "^2.4.3"
+    "body-parser": "^1.20.0",
+    "connect-flash": "^0.1.1",
+    "connect-mongo": "^4.6.0",
+    "cookie-parser": "^1.4.6",
+    "debug": "^4.3.4",
+    "ejs": "^3.1.6",
+    "express": "^4.17.3",
+    "express-partials": "^0.3.0",
+    "express-session": "^1.17.2",
+    "mongodb": "^4.4.1",
+    "morgan": "^1.10.0",
+    "serve-favicon": "^2.5.0"
   },
   "devDependencies": {
-    "eslint": "^4.18.2",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-standard": "^3.0.1"
+    "eslint": "^7.32.0",
+    "eslint-config-standard": "^16.0.3",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^5.2.0"
   }
 }
