@@ -1,21 +1,22 @@
-var express = require('express');
-var partials = require('express-partials');
-var path = require('path');
-var favicon = require('serve-favicon');
-var logger = require('morgan');
-var cookieParser = require('cookie-parser');
-var session = require('express-session');
-var bodyParser = require('body-parser');
-var MongoStore=require("connect-mongo")(session);
-var settings=require('./settings');
+import express from 'express'
+import partials from 'express-partials'
+import path from 'path'
+// import favicon from 'serve-favicon'
+import logger from 'morgan'
+import cookieParser from 'cookie-parser'
+import session from 'express-session'
+import bodyParser from 'body-parser'
+import MongoStore from 'connect-mongo'
+import flash from 'connect-flash'
 
-var routes = require('./routes/index');
-var users = require('./routes/users');
-var flash = require('connect-flash');
-var app = express();
+import settings from './settings.mjs'
+import routes from './routes/index.mjs'
+// import users from './routes/users.mjs'
+
+const app = express()
 
 // view engine setup
-app.set('views', path.join(__dirname, 'views'));
+app.set('views', path.join(path.resolve(), 'views'));
 app.set('view engine', 'ejs');
 
 // uncomment after placing your favicon in /public
@@ -24,24 +25,24 @@ app.use(logger('dev'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(express.static(path.join(__dirname, 'public')));
+app.use(express.static(path.join(path.resolve(), 'public')));
 app.use(partials());
 app.use(flash());
 
 app.use(session({
+  resave: true,
+  saveUninitialized: true,
   secret: settings.cookieSecret,
   key: settings.db,//cookie name
   cookie: {maxAge: 1000 * 60 * 60 * 24 * 30},//30 days
-  store: new MongoStore({
-    db: settings.db,
-    host: settings.host,
-    port: settings.port
+  store: MongoStore.create({
+    mongoUrl: `mongodb://${settings.host}/${settings.db}`
   })
 }));
 app.use(function(req, res, next){
   res.locals.user = req.session.user;
-  var error = req.flash('error');
-  var success = req.flash('success');
+  const error = req.flash('error')
+  const success = req.flash('success')
   res.locals.error = error.length ? error : null;
   res.locals.success = success.length ? success : null;
   next();
@@ -51,7 +52,7 @@ app.use('/', routes);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
-  var err = new Error('Not Found');
+  const err = new Error('Not Found')
   err.status = 404;
   next(err);
 });
@@ -80,5 +81,4 @@ app.use(function(err, req, res, next) {
   });
 });
 
-
-module.exports = app;
+export default app
