@@ -1,8 +1,9 @@
-let express = require('express')
+import express from 'express'
 let router = express.Router()
 
 /* GET users listing. */
 router.get('/', function(req, res, next) {
   res.send('respond with a resource')
 })
-module.exports = router
+
+export default router
